refactor: 감정칩 타입 공통화
