+++ features/record/model/useRecordFlow.ts
@@ -1,5 +1,5 @@
+import type { EmotionChip } from "@/shared/types";
 import { useState } from "react";
-import type { EmotionChip } from "./types";
 
 type Phase = "explore" | "write" | "complete";
