+++ features/record/ui/EmotionResult.tsx
@@ -1,6 +1,6 @@
-import type { EmotionChip } from "@/features/record/model/types";
 import * as S from "@/features/record/style/EmotionResult.styles";
 import RightIcon from "@/shared/assets/icons/right.svg";
+import type { EmotionChip } from "@/shared/types";
 import { Chip, IconButton } from "@/shared/ui";
 import type React from "react";
